WholeWords flag dances

+++ src/xpiks-qt/KeywordsPresets/presetkeywordsmodelconfig.cpp
@@ -56,11 +56,11 @@ namespace KeywordsPresets {
     }
 
     void PresetKeywordsModelConfig::saveFromModel(const std::vector<PresetModel *> &presets) {
-        size_t size = presets.size();
+        int size = (int)presets.size();
         LOG_INTEGR_TESTS_OR_DEBUG << size;
 
         m_PresetData.resize(size);
-        for (size_t i = 0; i < size; i++) {
+        for (int i = 0; i < size; i++) {
             auto *item = presets[i];
             auto &name = item->m_PresetName;
             auto &keywordsModel = item->m_KeywordsModel;
